staging/pkg/mod/golang.org/x/net@v0.0.0-20210520170846-37e1c6afe023/html/testdata/webkit/scriptdata01.dat

#data
FOO<script>'Hello'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'Hello'"
|     "BAR"

#data
FOO<script></script>BAR
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|     "BAR"

#data
FOO<script></script >BAR
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|     "BAR"

#data
FOO<script></script/>BAR
#errors
(1,3): expected-doctype-but-got-chars
(1,21): self-closing-flag-on-end-tag
#new-errors
(1:21) end-tag-with-trailing-solidus
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|     "BAR"

#data
FOO<script></script/ >BAR
#errors
(1,3): expected-doctype-but-got-chars
(1,20): unexpected-character-after-solidus-in-tag
#new-errors
(1:21) unexpected-solidus-in-tag
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|     "BAR"

#data
FOO<script type="text/plain"></scriptx>BAR
#errors
(1,3): expected-doctype-but-got-chars
(1,42): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       type="text/plain"
|       "</scriptx>BAR"

#data
FOO<script></script foo=">" dd>BAR
#errors
(1,3): expected-doctype-but-got-chars
(1,31): attributes-in-end-tag
#new-errors
(1:31) end-tag-with-attributes
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|     "BAR"

#data
FOO<script>'<'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<'"
|     "BAR"

#data
FOO<script>'<!'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!'"
|     "BAR"

#data
FOO<script>'<!-'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-'"
|     "BAR"

#data
FOO<script>'<!--'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!--'"
|     "BAR"

#data
FOO<script>'<!---'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!---'"
|     "BAR"

#data
FOO<script>'<!-->'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-->'"
|     "BAR"

#data
FOO<script>'<!-->'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-->'"
|     "BAR"

#data
FOO<script>'<!-- potato'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- potato'"
|     "BAR"

#data
FOO<script>'<!-- <sCrIpt'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt'"
|     "BAR"

#data
FOO<script type="text/plain">'<!-- <sCrIpt>'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
(1,56): expected-script-data-but-got-eof
(1,56): expected-named-closing-tag-but-got-eof
#new-errors
(1:57) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       type="text/plain"
|       "'<!-- <sCrIpt>'</script>BAR"

#data
FOO<script type="text/plain">'<!-- <sCrIpt> -'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
(1,58): expected-script-data-but-got-eof
(1,58): expected-named-closing-tag-but-got-eof
#new-errors
(1:59) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       type="text/plain"
|       "'<!-- <sCrIpt> -'</script>BAR"

#data
FOO<script type="text/plain">'<!-- <sCrIpt> --'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
(1,59): expected-script-data-but-got-eof
(1,59): expected-named-closing-tag-but-got-eof
#new-errors
(1:60) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       type="text/plain"
|       "'<!-- <sCrIpt> --'</script>BAR"

#data
FOO<script>'<!-- <sCrIpt> -->'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt> -->'"
|     "BAR"

#data
FOO<script type="text/plain">'<!-- <sCrIpt> --!>'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
(1,61): expected-script-data-but-got-eof
(1,61): expected-named-closing-tag-but-got-eof
#new-errors
(1:62) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       type="text/plain"
|       "'<!-- <sCrIpt> --!>'</script>BAR"

#data
FOO<script type="text/plain">'<!-- <sCrIpt> -- >'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
(1,61): expected-script-data-but-got-eof
(1,61): expected-named-closing-tag-but-got-eof
#new-errors
(1:62) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       type="text/plain"
|       "'<!-- <sCrIpt> -- >'</script>BAR"

#data
FOO<script type="text/plain">'<!-- <sCrIpt '</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
(1,56): expected-script-data-but-got-eof
(1,56): expected-named-closing-tag-but-got-eof
#new-errors
(1:57) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       type="text/plain"
|       "'<!-- <sCrIpt '</script>BAR"

#data
FOO<script type="text/plain">'<!-- <sCrIpt/'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
(1,56): expected-script-data-but-got-eof
(1,56): expected-named-closing-tag-but-got-eof
#new-errors
(1:57) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       type="text/plain"
|       "'<!-- <sCrIpt/'</script>BAR"

#data
FOO<script type="text/plain">'<!-- <sCrIpt\'</script>BAR
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       type="text/plain"
|       "'<!-- <sCrIpt\'"
|     "BAR"

#data
FOO<script type="text/plain">'<!-- <sCrIpt/'</script>BAR</script>QUX
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       type="text/plain"
|       "'<!-- <sCrIpt/'</script>BAR"
|     "QUX"

#data
FOO<script><!--<script>-></script>--></script>QUX
#errors
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "<!--<script>-></script>-->"
|     "QUX"
